requirements.txt
mysql-connector-python>=8.0.33
tkcalendar>=1.9.1

